gameai/requirements.txt
certifi==2020.6.20
cloudpickle==1.3.0
cycler==0.10.0
ffmpeg-python==0.2.0
future==0.18.2
gym==0.17.2
imageio==2.9.0
kiwisolver==1.2.0
lxml==4.5.2
-e git+https://github.com/martinballa/malmo@4b2951296a2687ff71d0ef81e2de59503f34d26d#egg=malmoenv&subdirectory=MalmoEnv
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.2
six==1.15.0
